This commit was generated by cvs2svn to compensate for changes in r16293, which included commits to RCS files with non-trunk default branches.
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@16294 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/arch/win32/intl.c
@@ -2704,7 +2704,7 @@ int length;
     if (!(text_cache[en_resource >> 7] & (1 << ((en_resource >> 4) & 7)))) {
         j = (en_resource >> 4) + 1;
         for (i = 0; i < countof(language_table); i++) {
-            hRes = FindResourceEx(NULL, RT_STRING, MAKEINTRESOURCE(j), MAKELANGID(windows_to_iso[i].windows_code, SUBLANG_NEUTRAL));
+            hRes = FindResourceEx(NULL, RT_STRING, MAKEINTRESOURCE(j), (WORD) MAKELANGID(windows_to_iso[i].windows_code, SUBLANG_NEUTRAL));
             if (hRes) {
                 hGlob = LoadResource(NULL, hRes);
                 p = LockResource(hGlob);
